hq playlist
it used to be that any items that were NOT in a playlist were
considered part of the HQ set.

This meant when I created a favourites playlist, a bunch of HQ episodes
went missing.

Now the name is parsed to see if it contains Boss HQ, but really I
should just make a HQ playlist on Youtube